+++ src/main/java/ch/sbb/polarion/extension/docx_exporter/util/LiveDocCommentsProcessor.java
@@ -1,5 +1,6 @@
 package ch.sbb.polarion.extension.docx_exporter.util;
 
+import ch.sbb.polarion.extension.docx_exporter.rest.model.conversion.CommentsRenderType;
 import ch.sbb.polarion.extension.generic.regex.RegexMatcher;
 import ch.sbb.polarion.extension.docx_exporter.model.LiveDocComment;
 import com.polarion.alm.server.api.model.document.ProxyDocument;
@@ -18,11 +19,13 @@ import com.polarion.alm.shared.api.model.user.UserField;
 import lombok.Builder;
 import lombok.Data;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.text.SimpleDateFormat;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
+import java.util.Optional;
 
 @SuppressWarnings("java:S1200")
 public class LiveDocCommentsProcessor {
@@ -76,19 +79,22 @@ public class LiveDocCommentsProcessor {
     }
 
     @NotNull
-    public String addLiveDocComments(ProxyDocument document, @NotNull String html, boolean onlyOpen) {
-        //Polarion document keeps comments position in spans marked with id 'polarion-comment:commentId'.
-        //<span id="polarion-comment:1"></span>"
-        //Following expression retrieves such spans.
-        final Map<String, LiveDocComment> liveDocComments = getCommentsFromDocument(document, onlyOpen);
-        return RegexMatcher.get("(?s)(<span id=\"polarion-comment:(?<commentId>\\d+)\"></span>)").replace(html, regexEngine -> {
-            String commentId = regexEngine.group("commentId");
-            LiveDocComment liveDocComment = liveDocComments.get(commentId);
-            return liveDocComment == null ? null : renderComment(liveDocComment, 0);
-        });
+    public String addLiveDocComments(ProxyDocument document, @NotNull String html, @Nullable CommentsRenderType commentsRenderType) {
+        //Polarion document keeps comments position in span/img marked with id 'polarion-comment:commentId'.
+        //<span id="polarion-comment:1"></span> or <img id="polarion-comment:1" class="polarion-dle-comment-icon"/> (for comments inside workitem description)
+        //Note that resolved comments have class 'polarion-dle-comment-resolved-icon' so we have to take care of them too.
+        final Map<String, LiveDocComment> liveDocComments = commentsRenderType == null ? Map.of() : getCommentsFromDocument(document, commentsRenderType.equals(CommentsRenderType.OPEN));
+        return RegexMatcher.get("(?s)((<img id=\"polarion-comment:(?<imgCommentId>\\d+)\"[^>]*class=\"polarion-dle-comment(?:-resolved)?-icon\"/>)|(<span id=\"polarion-comment:(?<spanCommentId>\\d+)\"></span>))")
+                .replace(html, regexEngine -> {
+                    String commentId = Optional.ofNullable(regexEngine.group("imgCommentId")).orElse(regexEngine.group("spanCommentId"));
+                    return renderComment(liveDocComments.get(commentId), 0);
+                });
     }
 
     private String renderComment(LiveDocComment liveDocComment, int nestingLevel) {
+        if (liveDocComment == null) {
+            return "";
+        }
         CommentData commentData = getCommentData(liveDocComment);
         StringBuilder commentSpan = new StringBuilder(getCommentSpan(commentData, nestingLevel));
         Map<String, LiveDocComment> childComments = liveDocComment.getChildComments();