+++ src/test/java/ch/sbb/polarion/extension/docx_exporter/pandoc/BasicTest.java
@@ -2,33 +2,26 @@ package ch.sbb.polarion.extension.docx_exporter.pandoc;
 
 import ch.sbb.polarion.extension.docx_exporter.pandoc.service.model.PandocInfo;
 import ch.sbb.polarion.extension.docx_exporter.pandoc.service.model.PandocParams;
-import ch.sbb.polarion.extension.docx_exporter.util.MediaUtils;
 import jakarta.xml.bind.JAXBElement;
 import lombok.SneakyThrows;
 import org.apache.pdfbox.Loader;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.docx4j.Docx4J;
 import org.docx4j.convert.out.FOSettings;
+import org.docx4j.fonts.PhysicalFonts;
 import org.docx4j.openpackaging.exceptions.Docx4JException;
 import org.docx4j.openpackaging.packages.WordprocessingMLPackage;
 import org.docx4j.openpackaging.parts.WordprocessingML.MainDocumentPart;
 import org.docx4j.wml.SectPr;
 import org.docx4j.wml.Text;
-import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import javax.imageio.ImageIO;
-import java.awt.*;
-import java.awt.image.BufferedImage;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -36,6 +29,15 @@ import static org.junit.jupiter.api.Assertions.*;
 @SkipTestWhenParamNotSet
 @SuppressWarnings("ResultOfMethodCallIgnored")
 class BasicTest extends BasePandocTest {
+
+    static {
+        // Workaround for docx4j font discovery issue with some OpenType fonts.
+        // Must be set BEFORE IdentityPlusMapper class is loaded (which triggers PhysicalFonts.discoverPhysicalFonts()).
+        // If this is not set early enough, docx4j may discover problematic OpenType fonts,
+        // which can cause exceptions or incorrect font rendering in generated documents, leading to test failures.
+        // Regex limits font discovery to common fonts, avoiding problematic OpenType fonts.
+        PhysicalFonts.setRegex(".*(Courier New|Arial|Times New Roman|Georgia|Trebuchet|Verdana|Helvetica).*");
+    }
 
     @Test
     void testInvalidTemplate() {
@@ -80,13 +82,6 @@ class BasicTest extends BasePandocTest {
         } finally {
             newFile.delete();
         }
-    }
-
-    private static @NotNull File getTestFile() throws IOException {
-        Path tempPath = Files.createTempFile("test", ".docx");
-        File newFile = tempPath.toFile();
-        newFile.deleteOnExit();
-        return newFile;
     }
 
     @Test
@@ -286,25 +281,6 @@ class BasicTest extends BasePandocTest {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         Docx4J.toFO(foSettings, baos, Docx4J.FLAG_EXPORT_PREFER_XSL);
         return baos.toByteArray();
-    }
-
-    private void compareContentUsingReferenceImages(String testName, byte[] pdf) throws IOException {
-        writeReportPdf(testName, "generated", pdf);
-        List<BufferedImage> resultImages = getAllPagesAsImagesAndLogAsReports(testName, pdf);
-        boolean hasDiff = false;
-        for (int i = 0; i < resultImages.size(); i++) {
-            try (InputStream inputStream = readPngResource(testName + PAGE_SUFFIX + i)) {
-                BufferedImage expectedImage = ImageIO.read(inputStream);
-                BufferedImage resultImage = resultImages.get(i);
-                List<Point> diffPoints = MediaUtils.diffImages(expectedImage, resultImage);
-                if (!diffPoints.isEmpty()) {
-                    MediaUtils.fillImagePoints(resultImage, diffPoints, Color.BLUE.getRGB());
-                    writeReportImage(String.format("%s%s%d_diff", testName, PAGE_SUFFIX, i), resultImage);
-                    hasDiff = true;
-                }
-            }
-        }
-        assertFalse(hasDiff);
     }
 
     private String generateLargeHtmlContent() {