+++ src/test/java/ch/sbb/polarion/extension/docx_exporter/util/LiveDocCommentsProcessorTest.java
@@ -1,5 +1,6 @@
 package ch.sbb.polarion.extension.docx_exporter.util;
 
+import ch.sbb.polarion.extension.docx_exporter.rest.model.conversion.CommentsRenderType;
 import com.polarion.alm.server.api.model.document.ProxyDocument;
 import com.polarion.alm.shared.api.model.comment.CommentBase;
 import com.polarion.alm.shared.api.model.comment.CommentBasesTreeField;
@@ -56,11 +57,11 @@ class LiveDocCommentsProcessorTest {
                 <div>some content1</div>
                 [span class=comment level-0][span class=meta][span class=date]2025-03-13 16:21[/span][span class=author]author1[/span][/span][span class=text]text1[/span][/span]
                 <div>some content2</div>
-                <span id="polarion-comment:2"></span>
+
                 <div>some content3</div>
                 [span class=comment level-0][span class=meta][span class=date]2025-03-13 16:23[/span][span class=author]author3[/span][/span][span class=text]text3[/span][/span]
                 <div>some content4</div>
-                """, new LiveDocCommentsProcessor().addLiveDocComments(document, html, true));
+                """, new LiveDocCommentsProcessor().addLiveDocComments(document, html, CommentsRenderType.OPEN));
 
         when(comments.iterator()).thenReturn(commentBases.iterator());
 
@@ -73,7 +74,7 @@ class LiveDocCommentsProcessorTest {
                 <div>some content3</div>
                 [span class=comment level-0][span class=meta][span class=date]2025-03-13 16:23[/span][span class=author]author3[/span][/span][span class=text]text3[/span][/span]
                 <div>some content4</div>
-                """, new LiveDocCommentsProcessor().addLiveDocComments(document, html, false));
+                """, new LiveDocCommentsProcessor().addLiveDocComments(document, html, CommentsRenderType.ALL));
     }
 
     @SneakyThrows